Start with framework detectors
